requirements.txt
django==3.2.10
environs==9.5.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
django-phonenumber-field[phonenumbers]==7.0.1
Pillow==9.3.0
requests==2.28.1
djangorestframework==3.14.0